Mattupham/validator table highlight row (#2289)
* add hover to table

* Fix row selection

* Update props

* Update with observer

* Add useDimensionStyle

* Enable useEffect

* Add cursor pointer

* Remove unused interface

* Clean up

* Remove td style

* Add container props
